# bpo-44850: Improve the performance of methodcaller.