--- a/Misc/NEWS.d/next/Library/2021-08-16-17-52-26.bpo-44850.r8jx5u.rst
+++ b/Misc/NEWS.d/next/Library/2021-08-16-17-52-26.bpo-44850.r8jx5u.rst
@@ -1,0 +1,2 @@
+Calls to ``operator.methodcaller`` are now 25-33% faster thanks to the use of
+the vectorcall protocol.