@sweeneyde 2021-08-17
A thought: there could be another performance boost in adding a tp_vectorcall to the `methodcaller` type, to avoid building the `args` tuple.