@sweeneyde 2021-08-18
Is that restriction still applicable since `Py_TPFLAGS_IMMUTABLETYPE` was added?