@neonene 2021-08-18
I think #27001 have made this possible.