@neonene 2021-08-18
The patch should work adding `Py_TPFLAGS_HAVE_VECTORCALL`  in `methodcaller_type_spec`.