@anntzer 2021-08-18
Ah yes, thanks, it now works, and this provided another ~25-33% speedup on the same benchmarks.