@markshannon 2021-08-19
Without benchmarks showing `methodcaller` being used in a larger program, you can't claim any sort of performance improvements.

This PR also makes `methodcaller` objects larger and slower to create. The microbenchmarks don't account for that.